+++ init.lua
@@ -38,7 +38,6 @@ function module:enable(config)
     recorder:prepareRecording()
     ui.resetButtons()
   end)
-  observe(native.addr(0x490690),8,function() recorder:feed() end)
   observe(native.addr(0x46b358),6,function(registers)
     if registers.EBP==61 and not engine.loading then recorder:reset(); ui.resetButtons() end
   end)